Beginning of overlaymanager

+++ Radiant/Radiant/CameraManager.cpp
@@ -31,8 +31,8 @@ const void CameraManager::CreateCamera(Entity entity)
 	Options* o = System::GetOptions();
 	CameraData cData(entity);
 	cData.aspect = static_cast<float>(System::GetWindowHandler()->GetWindowWidth()) / static_cast<float>(System::GetWindowHandler()->GetWindowHeight());
-	cData.fov = XMConvertToRadians( o->GetFoV() );
-	cData.farp = o->GetViewDistance();
+	cData.fov = (float)XMConvertToRadians( o->GetFoV() );
+	cData.farp = (float)o->GetViewDistance();
 	DirectX::XMStoreFloat4x4(&cData.projectionMatrix, DirectX::XMMatrixPerspectiveFovLH(cData.fov, cData.aspect, cData.nearp, cData.farp));
 
 	_entityToIndex[entity] = static_cast<int>(_cameras.size());
